requirements.txt
Django==1.6.5
Jinja2==2.7.3
Markdown==2.5.2
MySQL-python==1.2.3
### PyLEMS==0.4.3
PyNN==0.7.5
### PyOSB==0.3.0
PyYAML==3.11
Pygments==2.0.1
Sphinx==1.2.3
## FIXME: could not find svn URL in dependency_links for this package:
### airspeed==0.4.2dev-20141211
argh==0.26.1
argparse==1.2.1
backports.ssl-match-hostname==3.4.0.2
certifi==14.05.14
django-filter==0.9.1
git
-e git+https://github.com/danirus/django-pagedown.git@b67da0a5c68c8f1d75922afe7f5dc8eac3172a0b#egg=django_pagedown-master
djangorestframework==3.0.2
docutils==0.12
freemix==0.2
hidapi==0.7.99-5
impacket==0.9.11
jsonpickle==0.8.0
libNeuroML==0.2.4
livereload==2.3.2
mpi4py==1.3.1
pathtools==0.1.2
pp==1.5.7-pathos
processing==0.52-pathos
pymongo==2.7.2
pyramidal==0.1.0
sphinx-autobuild==0.3.0
symmetricjsonrpc==0.1.0
tornado==4.0.2
virtualenv==1.7.2
watchdog==0.8.2
wsgiref==0.1.2
zenmap==6.46
